requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests          # JSON‑RPC calls
google-adk        # ADK core
google-generativeai>=0.4.0